requirements.txt
crewai>=0.28.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
langchain>=0.1.0
langchain-community>=0.1.0
langchain-google-genai>=0.0.5
pydantic>=2.5.0
pyyaml>=6.0.0
chromadb>=0.4.18
pypdf>=3.15.1
requests>=2.31.0